Switch to target_sources and target_include_directories
Switch from setting XYCE_SOURCES and XYCE_INCLUDE_DIRECTORIES
to setting those values as properties of the XyceStatic target.

When they are consumed by the Xyce target, capture the data from the
target properties of XyceStatic.

Due to the setting of CMP0079, a CMAKE_CURRENT_SOURCE_DIR is no longer
necessary when adding sources as the relative path is assumed.